Firestore setup and create user
